+++ src/js/easymde.js
@@ -1,4 +1,3 @@
-/*global require,module*/
 'use strict';
 var CodeMirror = require('codemirror');
 require('codemirror/addon/edit/continuelist.js');
@@ -121,6 +120,11 @@ function createToolbarButton(options, enableTooltips, shortcuts) {
     el.className = options.name;
     el.setAttribute('type', 'button');
     enableTooltips = (enableTooltips == undefined) ? true : enableTooltips;
+
+    // Properly hande custom shortcuts
+    if (options.name && options.name in shortcuts) {
+        bindings[options.name] = options.action;
+    }
 
     if (options.title && enableTooltips) {
         el.title = createTooltip(options.title, options.action, shortcuts);
@@ -287,7 +291,7 @@ function toggleFullScreen(editor) {
     if (/editor-preview-active-side/.test(sidebyside.className))
         toggleSideBySide(editor);
 
-	if (editor.options.onToggleFullScreen) {
+    if (editor.options.onToggleFullScreen) {
         editor.options.onToggleFullScreen(cm.getOption('fullScreen') || false);
     }
 }
@@ -384,9 +388,9 @@ function toggleCodeBlock(editor) {
             line: start_line_sel,
             ch: 0,
         }, {
-            line: end_line_sel,
-            ch: 0,
-        });
+                line: end_line_sel,
+                ch: 0,
+            });
     }
 
     var cm = editor.codemirror,
@@ -408,9 +412,9 @@ function toggleCodeBlock(editor) {
             line: cur_start.line,
             ch: 0,
         }, {
-            line: cur_start.line,
-            ch: 99999999999999,
-        });
+                line: cur_start.line,
+                ch: 99999999999999,
+            });
         cur_start.ch--;
         if (cur_start !== cur_end) {
             cur_end.ch--;
@@ -469,24 +473,24 @@ function toggleCodeBlock(editor) {
                     line: end_line,
                     ch: 0,
                 }, {
-                    line: end_line + (end_text ? 0 : 1),
-                    ch: 0,
-                });
+                        line: end_line + (end_text ? 0 : 1),
+                        ch: 0,
+                    });
                 cm.replaceRange(start_text, {
                     line: start_line,
                     ch: 0,
                 }, {
-                    line: start_line + (start_text ? 0 : 1),
-                    ch: 0,
-                });
+                        line: start_line + (start_text ? 0 : 1),
+                        ch: 0,
+                    });
             });
             cm.setSelection({
                 line: start_line + (start_text ? 1 : 0),
                 ch: 0,
             }, {
-                line: end_line + (start_text ? 1 : -1),
-                ch: 0,
-            });
+                    line: end_line + (start_text ? 1 : -1),
+                    ch: 0,
+                });
             cm.focus();
         } else {
             // no selection, search for ends of this fenced block
@@ -522,16 +526,16 @@ function toggleCodeBlock(editor) {
                     line: block_start,
                     ch: 0,
                 }, {
-                    line: block_start + 1,
-                    ch: 0,
-                });
+                        line: block_start + 1,
+                        ch: 0,
+                    });
                 cm.replaceRange('', {
                     line: block_end - 1,
                     ch: 0,
                 }, {
-                    line: block_end,
-                    ch: 0,
-                });
+                        line: block_end,
+                        ch: 0,
+                    });
             });
             cm.focus();
         }
@@ -711,6 +715,33 @@ function drawImage(editor) {
 }
 
 /**
+ * Action for opening the browse-file window to upload an image to a server.
+ * @param editor {EasyMDE} The EasyMDE object
+ */
+function drawUploadedImage(editor) {
+    // TODO: Draw the image template with a fake url? ie: '![](importing foo.png...)'
+    editor.openBrowseFileWindow();
+}
+
+/**
+ * Action executed after an image have been successfully imported on the server.
+ * @param editor {EasyMDE} The EasyMDE object
+ * @param url {string} The url of the uploaded image
+ */
+function afterImageUploaded(editor, url) {
+    var cm = editor.codemirror;
+    var stat = getState(cm);
+    var options = editor.options;
+    var imageName = url.substr(url.lastIndexOf('/') + 1);
+    _replaceSelection(cm, stat.image, options.insertTexts.uploadedImage, url);
+    // show uploaded image filename for 1000ms
+    editor.updateStatusBar('upload-image', editor.options.imageTexts.sbOnUploaded.replace('#image_name#', imageName));
+    setTimeout(function () {
+        editor.updateStatusBar('upload-image', editor.options.imageTexts.sbInit);
+    }, 1000);
+}
+
+/**
  * Action for drawing a table.
  */
 function drawTable(editor) {
@@ -821,9 +852,23 @@ function togglePreview(editor) {
     var toolbar_div = wrapper.previousSibling;
     var toolbar = editor.options.toolbar ? editor.toolbarElements.preview : false;
     var preview = wrapper.lastChild;
-    if (!preview || !/editor-preview/.test(preview.className)) {
+    if (!preview || !/editor-preview-full/.test(preview.className)) {
+
         preview = document.createElement('div');
-        preview.className = 'editor-preview';
+        preview.className = 'editor-preview-full';
+
+        if (editor.options.previewClass) {
+
+            if (Array.isArray(editor.options.previewClass)) {
+                for (var i = 0; i < editor.options.previewClass.length; i++) {
+                    preview.className += (' ' + editor.options.previewClass[i]);
+                }
+
+            } else if (typeof editor.options.previewClass === 'string') {
+                preview.className += (' ' + editor.options.previewClass);
+            }
+        }
+
         wrapper.appendChild(preview);
     }
     if (/editor-preview-active/.test(preview.className)) {
@@ -866,6 +911,7 @@ function _replaceSelection(cm, active, startEnd, url) {
     Object.assign(startPoint, cm.getCursor('start'));
     Object.assign(endPoint, cm.getCursor('end'));
     if (url) {
+        start = start.replace('#url#', url);  // url is in start for upload-image
         end = end.replace('#url#', url);
     }
     if (active) {
@@ -951,9 +997,9 @@ function _toggleHeading(cm, direction, size) {
                 line: i,
                 ch: 0,
             }, {
-                line: i,
-                ch: 99999999999999,
-            });
+                    line: i,
+                    ch: 99999999999999,
+                });
         })(i);
     }
     cm.focus();
@@ -990,11 +1036,25 @@ function _toggleLine(cm, name) {
         var map = {
             'quote': '>',
             'unordered-list': '*',
-            'ordered-list': 'd+.',
+            'ordered-list': '\\d+.',
         };
         var rt = new RegExp(map[name]);
 
         return char && rt.test(char);
+    };
+
+    var _toggle = function (name, text, untoggleOnly) {
+        var arr = listRegexp.exec(text);
+        var char = _getChar(name, line);
+        if (arr !== null) {
+            if (_checkChar(name, arr[2])) {
+                char = '';
+            }
+            text = arr[1] + char + arr[3] + text.replace(whitespacesRegexp, '').replace(repl[name], '$1');
+        } else if (untoggleOnly == false) {
+            text = char + ' ' + text;
+        }
+        return text;
     };
 
     var line = 1;
@@ -1004,25 +1064,22 @@ function _toggleLine(cm, name) {
             if (stat[name]) {
                 text = text.replace(repl[name], '$1');
             } else {
-                var arr = listRegexp.exec(text);
-                var char = _getChar(name, line);
-                if (arr !== null) {
-                    if (_checkChar(name, arr[2])) {
-                        char = '';
-                    }
-                    text = arr[1] + char + arr[3] + text.replace(whitespacesRegexp, '').replace(repl[name], '$1');
-                } else {
-                    text = char + ' ' + text;
+                // If we're toggling unordered-list formatting, check if the current line
+                // is part of an ordered-list, and if so, untoggle that first.
+                // Workaround for https://github.com/Ionaru/easy-markdown-editor/issues/92
+                if (name == 'unordered-list') {
+                    text = _toggle('ordered-list', text, true);
                 }
+                text = _toggle(name, text, false);
                 line += 1;
             }
             cm.replaceRange(text, {
                 line: i,
                 ch: 0,
             }, {
-                line: i,
-                ch: 99999999999999,
-            });
+                    line: i,
+                    ch: 99999999999999,
+                });
         })(i);
     }
     cm.focus();
@@ -1061,9 +1118,9 @@ function _toggleBlock(editor, type, start_chars, end_chars) {
             line: startPoint.line,
             ch: 0,
         }, {
-            line: startPoint.line,
-            ch: 99999999999999,
-        });
+                line: startPoint.line,
+                ch: 99999999999999,
+            });
 
         if (type == 'bold' || type == 'strikethrough') {
             startPoint.ch -= 2;
@@ -1113,16 +1170,34 @@ function _cleanBlock(cm) {
             line: line,
             ch: 0,
         }, {
-            line: line,
-            ch: 99999999999999,
-        });
+                line: line,
+                ch: 99999999999999,
+            });
     }
+}
+
+/**
+ * Convert a number of bytes to a human-readable file size.
+ * @param bytes {integer} A number of bytes, as integer. Ex: 421137
+ * @param units {number[]} An array of human-readable units, ie. ['b', 'Kb', 'Mb']
+ * @returns string A human-readable file size. Ex: '412Kb'
+ */
+function humanFileSize(bytes, units) {
+    if (Math.abs(bytes) < 1024) {
+        return '' + bytes + units[0];
+    }
+    var u = 0;
+    do {
+        bytes /= 1024;
+        ++u;
+    } while (Math.abs(bytes) >= 1024 && u < units.length);
+    return '' + bytes.toFixed(1) + units[u];
 }
 
 // Merge the properties of one object into another.
 function _mergeProperties(target, source) {
     for (var property in source) {
-        if (source.hasOwnProperty(property)) {
+        if (Object.prototype.hasOwnProperty.call(source, property)) {
             if (source[property] instanceof Array) {
                 target[property] = source[property].concat(target[property] instanceof Array ? target[property] : []);
             } else if (
@@ -1276,6 +1351,12 @@ var toolbarBuiltInButtons = {
         title: 'Insert Image',
         default: true,
     },
+    'upload-image': {
+        name: 'upload-image',
+        action: drawUploadedImage,
+        className: 'fa fa-image',
+        title: 'Import an image',
+    },
     'table': {
         name: 'table',
         action: drawTable,
@@ -1350,6 +1431,8 @@ var toolbarBuiltInButtons = {
 var insertTexts = {
     link: ['[', '](#url#)'],
     image: ['![](', '#url#)'],
+    uploadedImage: ['![](#url#)', ''],
+    // uploadedImage: ['![](#url#)\n', ''], // TODO: New line insertion doesn't work here.
     table: ['', '\n\n| Column 1 | Column 2 | Column 3 |\n| -------- | -------- | -------- |\n| Text     | Text     | Text     |\n\n'],
     horizontalRule: ['', '\n\n-----\n\n'],
 };
@@ -1363,6 +1446,31 @@ var blockStyles = {
     'bold': '**',
     'code': '```',
     'italic': '*',
+};
+
+/**
+ * Texts displayed to the user (mainly on the status bar) for the import image
+ * feature. Can be used for customization or internationalization.
+ */
+var imageTexts = {
+    sbInit: 'Attach files by drag and dropping or pasting from clipboard.',
+    sbOnDragEnter: 'Drop image to upload it.',
+    sbOnDrop: 'Uploading image #images_names#...',
+    sbProgress: 'Uploading #file_name#: #progress#%',
+    sbOnUploaded: 'Uploaded #image_name#',
+    sizeUnits: 'b,Kb,Mb',
+};
+
+/**
+ * Errors displayed to the user, using the `errorCallback` option. Can be used for
+ * customization or internationalization.
+ */
+var errorMessages = {
+    noFileGiven: 'You must select a file.',
+    typeNotAllowed: 'This image type is not allowed.',
+    fileTooLarge: 'Image #image_name# is too big (#image_size#).\n' +
+        'Maximum file size is #image_max_size#.',
+    importError: 'Something went wrong when uploading the image #image_name#.',
 };
 
 /**
@@ -1420,7 +1528,7 @@ function EasyMDE(options) {
 
         // Loop over the built in buttons, to get the preferred order
         for (var key in toolbarBuiltInButtons) {
-            if (toolbarBuiltInButtons.hasOwnProperty(key)) {
+            if (Object.prototype.hasOwnProperty.call(toolbarBuiltInButtons, key)) {
                 if (key.indexOf('separator-') != -1) {
                     options.toolbar.push('|');
                 }
@@ -1432,10 +1540,18 @@ function EasyMDE(options) {
         }
     }
 
+    // Editor preview styling class.
+    if (!Object.prototype.hasOwnProperty.call(options, 'previewClass')) {
+        options.previewClass = 'editor-preview';
+    }
 
     // Handle status bar
-    if (!options.hasOwnProperty('status')) {
+    if (!Object.prototype.hasOwnProperty.call(options, 'status')) {
         options.status = ['autosave', 'lines', 'words', 'cursor'];
+
+        if (options.uploadImage) {
+            options.status.unshift('upload-image');
+        }
     }
 
 
@@ -1471,6 +1587,17 @@ function EasyMDE(options) {
 
     options.minHeight = options.minHeight || '300px';
 
+    options.errorCallback = options.errorCallback || function (errorMessage) {
+        alert(errorMessage);
+    };
+
+    // Import-image default configuration
+    options.uploadImage = options.uploadImage || false;
+    options.imageMaxSize = options.imageMaxSize || 2097152; // 1024 * 1024 * 2
+    options.imageAccept = options.imageAccept || 'image/png, image/jpeg';
+    options.imageTexts = extend({}, imageTexts, options.imageTexts || {});
+    options.errorMessages = extend({}, errorMessages, options.errorMessages || {});
+
 
     // Change unique_id to uniqueId for backwards compatibility
     if (options.autosave != undefined && options.autosave.unique_id != undefined && options.autosave.unique_id != '')
@@ -1491,7 +1618,106 @@ function EasyMDE(options) {
     if (options.initialValue && (!this.options.autosave || this.options.autosave.foundSavedValue !== true)) {
         this.value(options.initialValue);
     }
+
+    if (options.uploadImage) {
+        var self = this;
+
+        this.codemirror.on('dragenter', function (cm, event) {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbOnDragEnter);
+            event.stopPropagation();
+            event.preventDefault();
+        });
+        this.codemirror.on('dragend', function (cm, event) {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbInit);
+            event.stopPropagation();
+            event.preventDefault();
+        });
+        this.codemirror.on('dragleave', function (cm, event) {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbInit);
+            event.stopPropagation();
+            event.preventDefault();
+        });
+
+        this.codemirror.on('dragover', function (cm, event) {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbOnDragEnter);
+            event.stopPropagation();
+            event.preventDefault();
+        });
+
+        this.codemirror.on('drop', function (cm, event) {
+            event.stopPropagation();
+            event.preventDefault();
+            if (options.imageUploadFunction) {
+                self.uploadImagesUsingCustomFunction(options.imageUploadFunction, event.dataTransfer.files);
+            } else {
+                self.uploadImages(event.dataTransfer.files);
+            }
+        });
+
+        this.codemirror.on('paste', function (cm, event) {
+            if (options.imageUploadFunction) {
+                self.uploadImagesUsingCustomFunction(options.imageUploadFunction, event.clipboardData.files);
+            } else {
+                self.uploadImages(event.clipboardData.files);
+            }
+        });
+    }
 }
+
+/**
+ * Upload asynchronously a list of images to a server.
+ *
+ * Can be triggered by:
+ * - drag&drop;
+ * - copy-paste;
+ * - the browse-file window (opened when the user clicks on the *upload-image* icon).
+ * @param {FileList} files The files to upload the the server.
+ * @param [onSuccess] {function} see EasyMDE.prototype.uploadImage
+ * @param [onError] {function} see EasyMDE.prototype.uploadImage
+ */
+EasyMDE.prototype.uploadImages = function (files, onSuccess, onError) {
+    var names = [];
+    for (var i = 0; i < files.length; i++) {
+        names.push(files[i].name);
+        this.uploadImage(files[i], onSuccess, onError);
+    }
+    this.updateStatusBar('upload-image', this.options.imageTexts.sbOnDrop.replace('#images_names#', names.join(', ')));
+};
+
+/**
+ * Upload asynchronously a list of images to a server.
+ *
+ * Can be triggered by:
+ * - drag&drop;
+ * - copy-paste;
+ * - the browse-file window (opened when the user clicks on the *upload-image* icon).
+ * @param imageUploadFunction {Function} The custom function to upload the image passed in options.
+ * @param {FileList} files The files to upload the the server.
+ */
+EasyMDE.prototype.uploadImagesUsingCustomFunction = function (imageUploadFunction, files) {
+    var names = [];
+    for (var i = 0; i < files.length; i++) {
+        names.push(files[i].name);
+        this.uploadImageUsingCustomFunction(imageUploadFunction, files[i]);
+    }
+    this.updateStatusBar('upload-image', this.options.imageTexts.sbOnDrop.replace('#images_names#', names.join(', ')));
+};
+
+/**
+ * Update an item in the status bar.
+ * @param itemName {string} The name of the item to update (ie. 'upload-image', 'autosave', etc.).
+ * @param content {string} the new content of the item to write in the status bar.
+ */
+EasyMDE.prototype.updateStatusBar = function (itemName, content) {
+    var matchingClasses = this.gui.statusbar.getElementsByClassName(itemName);
+    if (matchingClasses.length === 1) {
+        this.gui.statusbar.getElementsByClassName(itemName)[0].textContent = content;
+    } else if (matchingClasses.length === 0) {
+        console.log('EasyMDE: status bar item ' + itemName + ' was not found.');
+    } else {
+        console.log('EasyMDE: Several status bar items named ' + itemName + ' was found.');
+    }
+};
 
 /**
  * Default markdown render.
@@ -1563,7 +1789,12 @@ EasyMDE.prototype.render = function (el) {
         if (options.shortcuts[key] !== null && bindings[key] !== null) {
             (function (key) {
                 keyMaps[fixShortcut(options.shortcuts[key])] = function () {
-                    bindings[key](self);
+                    var action = bindings[key];
+                    if (typeof action === 'function') {
+                        action(self);
+                    } else if (typeof action === 'string') {
+                        window.open(action, '_blank');
+                    }
                 };
             })(key);
         }
@@ -1682,22 +1913,22 @@ EasyMDE.prototype.autosave = function () {
             return;
         }
 
-        if(this.options.autosave.binded !== true) {
-          if (easyMDE.element.form != null && easyMDE.element.form != undefined) {
-              easyMDE.element.form.addEventListener('submit', function () {
-                  clearTimeout(easyMDE.autosaveTimeoutId);
-                  easyMDE.autosaveTimeoutId = undefined;
+        if (this.options.autosave.binded !== true) {
+            if (easyMDE.element.form != null && easyMDE.element.form != undefined) {
+                easyMDE.element.form.addEventListener('submit', function () {
+                    clearTimeout(easyMDE.autosaveTimeoutId);
+                    easyMDE.autosaveTimeoutId = undefined;
 
-                  localStorage.removeItem('smde_' + easyMDE.options.autosave.uniqueId);
+                    localStorage.removeItem('smde_' + easyMDE.options.autosave.uniqueId);
 
-                  // Restart autosaving in case the submit will be cancelled down the line
-                  setTimeout(function() {
-                    easyMDE.autosave();
-                  }, easyMDE.options.autosave.delay || 10000);
-              });
-          }
+                    // Restart autosaving in case the submit will be cancelled down the line
+                    setTimeout(function () {
+                        easyMDE.autosave();
+                    }, easyMDE.options.autosave.delay || 10000);
+                });
+            }
 
-          this.options.autosave.binded = true;
+            this.options.autosave.binded = true;
         }
 
         if (this.options.autosave.loaded !== true) {
@@ -1751,6 +1982,152 @@ EasyMDE.prototype.clearAutosavedValue = function () {
     }
 };
 
+/**
+ * Open the browse-file window to upload an image to a server.
+ * @param [onSuccess] {function} see EasyMDE.prototype.uploadImage
+ * @param [onError] {function} see EasyMDE.prototype.uploadImage
+ */
+EasyMDE.prototype.openBrowseFileWindow = function (onSuccess, onError) {
+    var self = this;
+    var imageInput = this.gui.toolbar.getElementsByClassName('imageInput')[0];
+    imageInput.click(); //dispatchEvent(new MouseEvent('click'));  // replaced with click() for IE11 compatibility.
+    function onChange(event) {
+        self.uploadImages(event.target.files, onSuccess, onError);
+        imageInput.removeEventListener('change', onChange);
+    }
+
+    imageInput.addEventListener('change', onChange);
+};
+
+/**
+ * Upload an image to the server.
+ *
+ * @param file {File} The image to upload, as a HTML5 File object (https://developer.mozilla.org/en-US/docs/Web/API/File)
+ * @param [onSuccess] {function} A callback function to execute after the image has been successfully uploaded, with one parameter:
+ * - url (string): The URL of the uploaded image.
+ * @param [onError] {function} A callback function to execute when the image upload fails, with one parameter:
+ * - error (string): the detailed error to display to the user (based on messages from options.errorMessages).
+ */
+EasyMDE.prototype.uploadImage = function (file, onSuccess, onError) {
+    var self = this;
+    onSuccess = onSuccess || function onSuccess(imageUrl) {
+        afterImageUploaded(self, imageUrl);
+    };
+
+    function onErrorSup(errorMessage) {
+        // show error on status bar and reset after 10000ms
+        self.updateStatusBar('upload-image', errorMessage);
+
+        setTimeout(function () {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbInit);
+        }, 10000);
+
+        // run custom error handler
+        if (onError && typeof onError === 'function') {
+            onError(errorMessage);
+        }
+        // run error handler from options, this alerts the message.
+        self.options.errorCallback(errorMessage);
+    }
+
+    function fillErrorMessage(errorMessage) {
+        var units = self.options.imageTexts.sizeUnits.split(',');
+        return errorMessage
+            .replace('#image_name#', file.name)
+            .replace('#image_size#', humanFileSize(file.size, units))
+            .replace('#image_max_size#', humanFileSize(self.options.imageMaxSize, units));
+    }
+
+    if (file.size > this.options.imageMaxSize) {
+        onErrorSup(fillErrorMessage(this.options.errorMessages.fileTooLarge));
+        return;
+    }
+
+    var formData = new FormData();
+    formData.append('image', file);
+
+    // insert CSRF token if provided in config.
+    if (self.options.imageCSRFToken) {
+        formData.append('csrfmiddlewaretoken', self.options.imageCSRFToken);
+    }
+    var request = new XMLHttpRequest();
+    request.upload.onprogress = function (event) {
+        if (event.lengthComputable) {
+            var progress = '' + Math.round((event.loaded * 100) / event.total);
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbProgress.replace('#file_name#', file.name).replace('#progress#', progress));
+        }
+    };
+    request.open('POST', this.options.imageUploadEndpoint);
+
+    request.onload = function () {
+        try {
+            var response = JSON.parse(this.responseText);
+        } catch (error) {
+            console.error('EasyMDE: The server did not return a valid json.');
+            onErrorSup(fillErrorMessage(self.options.errorMessages.importError));
+            return;
+        }
+        if (this.status === 200 && response && !response.error && response.data && response.data.filePath) {
+            onSuccess(window.location.origin + '/' + response.data.filePath);
+        } else {
+            if (response.error && response.error in self.options.errorMessages) {  // preformatted error message
+                onErrorSup(fillErrorMessage(self.options.errorMessages[response.error]));
+            } else if (response.error) {  // server side generated error message
+                onErrorSup(fillErrorMessage(response.error));
+            } else {  //unknown error
+                console.error('EasyMDE: Received an unexpected response after uploading the image.'
+                    + this.status + ' (' + this.statusText + ')');
+                onErrorSup(fillErrorMessage(self.options.errorMessages.importError));
+            }
+        }
+    };
+
+    request.onerror = function (event) {
+        console.error('EasyMDE: An unexpected error occurred when trying to upload the image.'
+            + event.target.status + ' (' + event.target.statusText + ')');
+        onErrorSup(self.options.errorMessages.importError);
+    };
+
+    request.send(formData);
+
+};
+
+/**
+ * Upload an image to the server using a custom upload function.
+ *
+ * @param imageUploadFunction {Function} The custom function to upload the image passed in options
+ * @param file {File} The image to upload, as a HTML5 File object (https://developer.mozilla.org/en-US/docs/Web/API/File).
+ */
+EasyMDE.prototype.uploadImageUsingCustomFunction = function (imageUploadFunction, file) {
+    var self = this;
+    function onSuccess(imageUrl) {
+        afterImageUploaded(self, imageUrl);
+    }
+
+    function onError(errorMessage) {
+        var filledErrorMessage = fillErrorMessage(errorMessage);
+        // show error on status bar and reset after 10000ms
+        self.updateStatusBar('upload-image', filledErrorMessage);
+
+        setTimeout(function () {
+            self.updateStatusBar('upload-image', self.options.imageTexts.sbInit);
+        }, 10000);
+
+        // run error handler from options, this alerts the message.
+        self.options.errorCallback(filledErrorMessage);
+    }
+
+    function fillErrorMessage(errorMessage) {
+        var units = self.options.imageTexts.sizeUnits.split(',');
+        return errorMessage
+            .replace('#image_name#', file.name)
+            .replace('#image_size#', humanFileSize(file.size, units))
+            .replace('#image_max_size#', humanFileSize(self.options.imageMaxSize, units));
+    }
+
+    imageUploadFunction(file, onSuccess, onError);
+};
+
 EasyMDE.prototype.createSideBySide = function () {
     var cm = this.codemirror;
     var wrapper = cm.getWrapperElement();
@@ -1759,6 +2136,19 @@ EasyMDE.prototype.createSideBySide = function () {
     if (!preview || !/editor-preview-side/.test(preview.className)) {
         preview = document.createElement('div');
         preview.className = 'editor-preview-side';
+
+        if (this.options.previewClass) {
+
+            if (Array.isArray(this.options.previewClass)) {
+                for (var i = 0; i < this.options.previewClass.length; i++) {
+                    preview.className += (' ' + this.options.previewClass[i]);
+                }
+
+            } else if (typeof this.options.previewClass === 'string') {
+                preview.className += (' ' + this.options.previewClass);
+            }
+        }
+
         wrapper.parentNode.insertBefore(preview, wrapper.nextSibling);
     }
 
@@ -1869,6 +2259,20 @@ EasyMDE.prototype.createToolbar = function (items) {
 
             toolbarData[item.name || item] = el;
             bar.appendChild(el);
+
+            // Create the input element (ie. <input type='file'>), used among
+            // with the 'import-image' icon to open the browse-file window.
+            if (item.name === 'upload-image') {
+                var imageInput = document.createElement('input');
+                imageInput.className = 'imageInput';
+                imageInput.type = 'file';
+                imageInput.multiple = true;
+                imageInput.name = 'image';
+                imageInput.accept = self.options.imageAccept;
+                imageInput.style.display = 'none';
+                imageInput.style.opacity = 0;
+                bar.appendChild(imageInput);
+            }
         })(items[i]);
     }
 
@@ -1911,11 +2315,10 @@ EasyMDE.prototype.createStatusbar = function (status) {
     var options = this.options;
     var cm = this.codemirror;
 
-
     // Make sure the status variable is valid
-    if (!status || status.length === 0)
+    if (!status || status.length === 0) {
         return;
-
+    }
 
     // Set up the built-in items
     var items = [];
@@ -1964,6 +2367,10 @@ EasyMDE.prototype.createStatusbar = function (status) {
                     if (options.autosave != undefined && options.autosave.enabled === true) {
                         el.setAttribute('id', 'autosaved');
                     }
+                };
+            } else if (name === 'upload-image') {
+                defaultValue = function (el) {
+                    el.innerHTML = options.imageTexts.sbInit;
                 };
             }