+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # EasyMDE Changelog
-All notable changes to this project will be documented in this file.
+All notable changes to easymde will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Upload images functionality (Thanks to [@roipoussiere] and [@JeroenvO], [#71], [#101]).
+- Allow custom image upload function (Thanks to [@sperezp], [#106]).
+
+## [2.7.0] - 2019-07-13
+### Added
+- `previewClass` option for overwriting the preview screen class ([#99]).
+
+### Fixed
+- Updated dependencies to resolve potential security issue.
+- Resolved small code style issues shown by new eslint rules.
+
+## [2.6.1] - 2019-06-17
+### Fixed
+- Error when toggling between ordered and unordered lists (Thanks to [@roryok], [#93]).
+- Keyboard shortcuts for custom actions not working (Thanks to [@ysykzheng], [#75]).
+
+## [2.6.0] - 2019-04-15
 ### Added
 - Contributing guide (Thanks to [@roipoussiere], [#54]).
 - Issue templates.
@@ -13,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Finish rewrite of README (Thanks to [@roipoussiere], [#54]).
 - Image and link prompt fill with "https://" by default.
-- Link to markdown guide to https://www.markdownguide.org/basic-syntax/.
+- Link to markdown guide to <https://www.markdownguide.org/basic-syntax/>.
 
 ### Fixed
 - Backwards compatibility in the API with SimpleMDE 1.0.0 ([#41]).
@@ -95,6 +113,7 @@ Project forked from [SimpleMDE](https://github.com/sparksuite/simplemde-markdown
 - Cursor not always showing in "text" mode over the edit field
 
 <!-- Linked issues -->
+[#99]: https://github.com/Ionaru/easy-markdown-editor/issues/99
 [#45]: https://github.com/Ionaru/easy-markdown-editor/issues/45
 [#44]: https://github.com/Ionaru/easy-markdown-editor/issues/44
 [#41]: https://github.com/Ionaru/easy-markdown-editor/issues/41
@@ -106,12 +125,22 @@ Project forked from [SimpleMDE](https://github.com/sparksuite/simplemde-markdown
 [#9]: https://github.com/Ionaru/easy-markdown-editor/issues/9
 
 <!-- Linked PRs -->
+[#106]: https://github.com/Ionaru/easy-markdown-editor/pull/106
+[#101]: https://github.com/Ionaru/easy-markdown-editor/pull/101
+[#93]: https://github.com/Ionaru/easy-markdown-editor/pull/93
+[#75]: https://github.com/Ionaru/easy-markdown-editor/pull/75
+[#71]: https://github.com/Ionaru/easy-markdown-editor/pull/71
 [#54]: https://github.com/Ionaru/easy-markdown-editor/pull/54
 [#31]: https://github.com/Ionaru/easy-markdown-editor/pull/31
 [#27]: https://github.com/Ionaru/easy-markdown-editor/pull/27
 [#19]: https://github.com/Ionaru/easy-markdown-editor/pull/19
 
 <!-- Linked users -->
+[@sperezp]: https://github.com/sperezp
+[@JeroenvO]: https://github.com/JeroenvO
+[@sn3p]: https://github.com/sn3p
+[@roryok]: https://github.com/roryok
+[@ysykzheng]: https://github.com/ysykzheng
 [@roipoussiere]: https://github.com/roipoussiere
 [@FranklinWhale]: https://github.com/FranklinWhale
 [@Furgas]: https://github.com/Furgas
@@ -123,7 +152,10 @@ Project forked from [SimpleMDE](https://github.com/sparksuite/simplemde-markdown
 [@sne11ius]: https://github.com/sne11ius
 
 <!-- Linked versions -->
-[Unreleased]: https://github.com/Ionaru/easy-markdown-editor/compare/2.5.1...HEAD
+[Unreleased]: https://github.com/Ionaru/easy-markdown-editor/compare/2.7.0...HEAD
+[2.7.0]: https://github.com/Ionaru/easy-markdown-editor/compare/2.6.1...2.7.0
+[2.6.1]: https://github.com/Ionaru/easy-markdown-editor/compare/2.6.0...2.6.1
+[2.6.0]: https://github.com/Ionaru/easy-markdown-editor/compare/2.5.1...2.6.0
 [2.5.1]: https://github.com/Ionaru/easy-markdown-editor/compare/2.5.0...2.5.1
 [2.5.0]: https://github.com/Ionaru/easy-markdown-editor/compare/2.4.2...2.5.0
 [2.4.2]: https://github.com/Ionaru/easy-markdown-editor/compare/2.4.1...2.4.2